i3bar: allow "power savings" feature to be disabled
i3bar has a "power savings" feature that pauses the bar when it is hidden or
obscured by a fullscreen app. However this becomes a problem for some users who
use status line generators such as i3status-rust that have blocks that need
to run at all times. A hacky way to prevent this from happening is to set
stop_signal to a signal that won't cause the status line generator program to stop,
however I do not believe that is a good solution. This PR adds an option on the i3 side
to allow the user to disable this feature in the first place.

This has surprised users a fair amount of time over the past few years, to show a few:
https://github.com/greshake/i3status-rust/issues/503
https://github.com/greshake/i3status-rust/pull/694#issuecomment-630985696
https://faq.i3wm.org/question/4631/dont-sigstop-when-in-hide-mode.1.html
https://github.com/i3/i3/issues/3242#issuecomment-554636654
https://www.reddit.com/r/i3wm/comments/gm6yfh/is_there_a_painless_way_to_keep_i3bar_children/